+ 增加崩溃信息打印

+++ src/common/buffer_guard.cpp
@@ -54,7 +54,7 @@ void celeritas::buffer_guard::set_effective_size(size_t size)
     effective_size_ = size;
 }
 
-char* celeritas::buffer_guard::get(const int offset)
+char* celeritas::buffer_guard::get(const size_t offset)
 {
     return buffer_data_.get(offset);
 }
