+++ src/common/buffer_pool_data.cpp
@@ -50,7 +50,7 @@ bool celeritas::buffer_pool_data::is_effective() const noexcept
     return data_ != nullptr;
 }
 
-char* celeritas::buffer_pool_data::get(int offset)
+char* celeritas::buffer_pool_data::get(size_t offset)
 {
     if (is_effective())
     {
